develop

+++ tomcat.py
@@ -1,5 +1,6 @@
 import discord
 
+
 from src.constant import PATH_DATA
 from src.constant import PATH_IMG
 from src.constant import PATH_SCRIPT
@@ -7,7 +8,9 @@ from src.constant import PATH_SRC
 from src.constant import PATH_TEMP
 from src.constant import TOKEN
 
+import os.path
 import re
+import requests
 import random
 import subprocess
 from urllib.request import urlopen
@@ -15,7 +18,7 @@ from urllib.request import urlopen
 
 
 def run_command(command):
-    return subprocess.run(command.split(), stdout=subprocess.PIPE).stdout.decode()
+    return subprocess.run(command.split(), capture_output=True, text=True, universal_newlines=True).stdout
 
 
 def run_script(script, params=""):
@@ -34,8 +37,13 @@ def get_user(self, name):
 
 
 async def get_audio(self, message):
+    if len(message.content.split()) < 3:
+        await message.channel.send("Error: No se pasó ningún link para descargar. Try **tom --help**.")
     address = message.content.split()[2]
     file = run_script("download_audio.sh", address)
+    if file.startswith("Error"):
+        await message.channel.send(file)
+        return
     await message.channel.send(file=discord.File(file))
     run_command("rm " + file)
     return
@@ -58,7 +66,7 @@ async def cara_cruz(self, message):
 
 async def frieren(self, message):
     pass
-    
+ 
 
 async def get_dolar(self, message):
     url = "https://dolarhoy.com/cotizaciondolarblue"
@@ -96,23 +104,31 @@ async def message_to(self, message):
 
 
 async def teo_status(self, message):
-    run_script("teostatus.sh")
-    await message.channel.send(file=discord.File(PATH_TEMP + "img.png"))
-#    await message.channel.send(file=discord.File(PATH_IMG + "idk.png"))
+#    run_script("teostatus.sh")
+#    await message.channel.send(file=discord.File(PATH_TEMP + "img.png"))
+    await message.channel.send(file=discord.File(PATH_IMG + "idk.png"))
     return
 
 
+async def translate(self, message):
+    text = ' '.join(message.content.split()[2:])
+    text = "\""+text+"\""
+    await message.channel.send(run_script("translate.sh", text))
+
+
 async def raw_run(self, message):
-    if message.author.name == 'teo.730':
+    if message.author.id == 305483295833980938:
         command = ' '.join(message.content.split()[1:])
         response = run_command(command)
         if response:
-            await message.author.send(response)
+            await message.channel.send(response)
         return
 
 
 async def not_command(self, message):
-    await message.channel.send("El comando está **obsoleto**, consulta la documentación \"tom --help\" para aprender a usar correctamente el bot.")
+    await message.channel.send("Usa **\"tom --help\"** para ver comandos y ejemplos.")
+
+
 
 
 
@@ -124,17 +140,20 @@ commands = {
     "tom --cara-o-cruz": [cara_cruz],
     "tom -d": [get_dolar],
     "tom --dolar": [get_dolar],
-    "tom -f": [frieren],
-    "tom --frieren": [frieren],
+#    "tom -f": [frieren],
+#    "tom --frieren": [frieren],
     "tom -h": [print_help],
     "tom --help": [print_help],
     "tom -m": [message_to],
     "tom --mensaje": [message_to],
     "tom -q": [teo_status],
     "tom --que-hace-teo?": [teo_status],
-#    "tom -r": responde,
-#    "tom --responde": responde,
+    "tom -t": [translate],
+    "tom --traducir": [translate],
     "exec": [raw_run],
+
+    
+# Old commands
     "que hace teo?": [not_command,teo_status],
     "cara o ceca": [not_command, cara_cruz],
     "cara o cruz": [not_command, cara_cruz],
@@ -142,10 +161,11 @@ commands = {
     "tvd": [not_command, get_dolar],
 }
 
+
 class TomCat(discord.Client):
 
     async def on_ready(self):
-        print(f'Logged on as {self.user}!')
+        print(f'A darle atomos! {self.user}')
 
     async def on_message(self, message):
         if message.author == self.user:
@@ -154,15 +174,8 @@ class TomCat(discord.Client):
         for prefix, functions in commands.items():
             if message.content.lower().startswith(prefix):
                 for function in functions:
-                    print(message.content)
                     await function(self, message)
                 break
-
-    async def on_voice_state_update(self, user, before, after):
-        guild = self.get_guild(289528008027406337)
-        if user.id == 496099123695583253 and (before.channel == None or before.channel != after.channel) and after.self_mute and after.channel != None:
-            print("deleted juampa")
-            await user.move_to(discord.utils.get(guild.voice_channels, name='AFK'))
